api_gateway/requirements.txt
azure-functions
requests
azure-eventgrid
azure-core
requests
PyJWT